Move integration tests into their own folder
