+++ integration_test/ts/_proto/improbable/grpcweb/test/test_pb_service.d.ts
